refactor: move common utils code with openApi (parser) to generator/utils
