+++ packages/openapi-ts/src/generate/types.ts
@@ -6,14 +6,14 @@ import {
   type Node,
   TypeScriptFile,
 } from '../compiler';
+import { enumEntry, enumUnionType } from '../generator/utils/enum';
+import { escapeComment } from '../generator/utils/escape';
 import type { Model, OperationParameter } from '../openApi';
 import type { Method } from '../openApi';
 import { isOperationParameterRequired } from '../openApi/common/parser/operation';
 import type { Client } from '../types/client';
 import type { Files } from '../types/utils';
 import { getConfig, isStandaloneClient } from '../utils/config';
-import { enumEntry, enumUnionType } from '../utils/enum';
-import { escapeComment } from '../utils/escape';
 import { sortByName, sorterByName } from '../utils/sort';
 import {
   setUniqueTypeName,
