+++ packages/openapi-ts/src/generator/utils/__tests__/enum.spec.ts
