+++ packages/openapi-ts/src/generator/utils/__tests__/escape.spec.ts
