+++ packages/openapi-ts/src/generator/utils/__tests__/sanitize.spec.ts
