+++ packages/openapi-ts/src/generator/utils/camelCase.ts
