+++ packages/openapi-ts/src/generator/utils/const.ts
