+++ packages/openapi-ts/src/generator/utils/escape.ts
