+++ packages/openapi-ts/src/generator/utils/reservedWords.ts
