+++ packages/openapi-ts/src/generator/utils/unique.ts
