+++ packages/openapi-ts/tsconfig.json
@@ -7,5 +7,10 @@
   },
   "exclude": ["node_modules", "**/__mocks__"],
   "files": ["./src/types/hbs.d.ts"],
-  "include": ["./src/**/*.ts", "rollup.config.ts", "rollup.dts.config.ts"]
+  "include": [
+    "./src/**/*.ts",
+    "rollup.config.ts",
+    "rollup.dts.config.ts",
+    "src/generator/utils/const.ts"
+  ]
 }
